refactor: split CSVImporter into focused per-concern files (#80)
CSVImporter.swift packed ~20 types — row models, errors, format detection,
column mapping, date parsing, unit resolution, and the importer itself — into
one 1063-line file. Decomposed along the natural concern boundaries (the
existing MARK sections) into a sibling CSV/ folder:

  Baseline/Utilities/CSV/
    CSVTypes.swift          — public types (rows, results, errors,
                              CSVFormat + its format-detection extension,
                              ConflictStrategy, ParsedImport, ImportOutcome)
    CSVColumnMapping.swift  — ColumnRole, ColumnSynonyms, NormalizedHeader,
                              HeaderMap
    CSVDateParser.swift     — FlexibleDateParser
    CSVUnitResolvers.swift  — WeightUnitResolver, LengthUnitResolver
    CSVStringHelpers.swift  — String.stripBOM (was private to CSVImporter;
                              now shared internal so CSVFormat.detect can
                              call it from its new home)

CSVImporter.swift now contains just the importer enum: 1063 → 577 lines, out
of the oversized list. No behavior change.

Verified: make lint (0, --strict), make build, make test (272 logic incl.
~857 lines of CSVImporterTests + 13 UI, all green).
