+++ Baseline/Utilities/CSVImporter.swift
@@ -1,504 +1,6 @@
 import Foundation
 import SwiftData
 
-// MARK: - Public Types
-
-/// Row-level parse problem. Returned alongside successful rows so the
-/// caller can surface them to the user without aborting the whole import.
-struct CSVParseIssue: Equatable {
-    let line: Int
-    let reason: String
-}
-
-/// Typed output of parsing one CSV blob.
-struct CSVParseResult<Row: Equatable>: Equatable {
-    let rows: [Row]
-    let issues: [CSVParseIssue]
-}
-
-struct CSVWeightRow: Equatable {
-    let date: Date
-    let weight: Double
-    let unit: String
-    let notes: String?
-}
-
-struct CSVMeasurementRow: Equatable {
-    let date: Date
-    let type: MeasurementType
-    let valueCm: Double
-    let notes: String?
-}
-
-struct CSVScanRow: Equatable {
-    let date: Date
-    let type: ScanType
-    let source: ScanSource
-    let payload: InBodyPayload
-}
-
-enum CSVImportError: Error, Equatable {
-    case emptyFile
-    /// Header doesn't supply the columns the target format needs. Each
-    /// string is the semantic name of a missing column (e.g. "date",
-    /// "weight"). Callers surface these to the user verbatim.
-    case missingRequiredColumns([String])
-}
-
-/// Semantic identity of the parsed rows. Determined from which column
-/// roles are present in the header — no more exact header-string match.
-enum CSVFormat: String {
-    case weights
-    case measurements
-    case scans
-}
-
-/// How to handle rows whose (type, day) collides with existing data.
-enum ConflictStrategy {
-    /// Leave existing data untouched; skip the imported row.
-    case skip
-    /// Delete existing data (and its HealthKit samples) and persist the imported row.
-    case overwrite
-}
-
-/// Union of parsed-result types. Lets the UI (and dispatch helpers) hold
-/// a single state slot regardless of which format the user picked.
-enum ParsedImport {
-    case weights(CSVParseResult<CSVWeightRow>)
-    case measurements(CSVParseResult<CSVMeasurementRow>)
-    case scans(CSVParseResult<CSVScanRow>)
-
-    var format: CSVFormat {
-        switch self {
-        case .weights: return .weights
-        case .measurements: return .measurements
-        case .scans: return .scans
-        }
-    }
-
-    var rowCount: Int {
-        switch self {
-        case .weights(let r): return r.rows.count
-        case .measurements(let r): return r.rows.count
-        case .scans(let r): return r.rows.count
-        }
-    }
-
-    var issues: [CSVParseIssue] {
-        switch self {
-        case .weights(let r): return r.issues
-        case .measurements(let r): return r.issues
-        case .scans(let r): return r.issues
-        }
-    }
-}
-
-/// Error for `CSVImporter.parseAny` — distinguishes "I couldn't detect the
-/// format" from the lower-level parse errors.
-enum CSVDispatchError: Error, Equatable {
-    case unknownFormat
-    case parseFailed(CSVImportError)
-}
-
-/// Per-format counts returned from `import*` functions so the UI can summarize.
-struct ImportOutcome: Equatable {
-    var inserted: Int = 0
-    var overwritten: Int = 0
-    var skipped: Int = 0
-    var failed: Int = 0
-
-    static var empty: ImportOutcome { ImportOutcome() }
-}
-
-// MARK: - Column roles
-//
-// The semantic dimensions a CSV row can carry, independent of what the
-// source spreadsheet happens to call them. Parsers extract data by role,
-// not by column index or header string — so a header of `Weight (lb)` and
-// one of `lb` resolve to the same thing.
-
-/// Semantic role of a single CSV column.
-enum ColumnRole: String, Hashable, CaseIterable {
-    // Shared across all formats
-    case date
-    case time
-    case notes
-
-    // Weight entries
-    case weight
-    case weightUnit
-
-    // Tape measurements
-    case measurementType
-    case measurementValue
-
-    // InBody scans
-    case scanType
-    case scanSource
-    case scanWeightKg
-    case scanSMM
-    case scanBFM
-    case scanPBF
-    case scanTBW
-    case scanBMI
-    case scanBMR
-}
-
-/// Registry of acceptable header names for each role. Matching is
-/// case-insensitive, trims whitespace, and ignores parenthesized hints
-/// (which are captured separately for unit resolution).
-///
-/// **Extension point:** to accept a new synonym for an existing role,
-/// add it to the set. To introduce a new role, add a `ColumnRole` case
-/// and a matching entry here — the parsers pick it up via `HeaderMap`
-/// without any orchestration change.
-enum ColumnSynonyms {
-    static let registry: [ColumnRole: Set<String>] = [
-        .date: ["date", "day", "timestamp", "datetime"],
-        .time: ["time", "time of day", "clock"],
-        .notes: ["notes", "note", "comment", "comments", "memo"],
-
-        .weight: ["weight", "mass", "lb", "lbs", "kg", "kgs", "pounds", "kilograms"],
-        .weightUnit: ["unit", "units"],
-
-        .measurementType: ["type", "measurement", "part", "site"],
-        .measurementValue: ["value", "valuecm", "valuein", "measurement value", "cm", "in", "inches", "centimeters"],
-
-        .scanType: ["type", "scan type"],
-        .scanSource: ["source", "scan source"],
-        .scanWeightKg: ["weightkg", "weight_kg"],
-        .scanSMM: ["skeletalmusclemasskg", "skeletal_muscle_mass_kg", "smm"],
-        .scanBFM: ["bodyfatmasskg", "body_fat_mass_kg", "bfm"],
-        .scanPBF: ["bodyfatpct", "body_fat_pct", "pbf"],
-        .scanTBW: ["totalbodywaterl", "total_body_water_l", "tbw"],
-        .scanBMI: ["bmi"],
-        .scanBMR: ["basalmetabolicrate", "basal_metabolic_rate", "bmr"]
-    ]
-
-    /// Returns every role whose synonym set contains this normalised
-    /// header. Multiple roles can match (e.g. `type` matches both
-    /// `.measurementType` and `.scanType`); format detection later
-    /// disambiguates based on which required-column-sets are satisfied.
-    static func roles(forNormalized header: String) -> Set<ColumnRole> {
-        var matches: Set<ColumnRole> = []
-        for (role, synonyms) in registry where synonyms.contains(header) {
-            matches.insert(role)
-        }
-        return matches
-    }
-}
-
-// MARK: - Header parsing
-
-/// One header cell's raw form broken into the parts each consumer needs:
-/// the synonym-match key (`normalized`) and the optional unit hint
-/// extracted from any parenthesized suffix (`"Weight (lb)"` → `"lb"`).
-struct NormalizedHeader: Equatable {
-    let raw: String
-    let normalized: String
-    let parentheticalHint: String?
-
-    init(_ raw: String) {
-        self.raw = raw
-        let trimmed = raw.trimmingCharacters(in: .whitespaces)
-
-        if let openIdx = trimmed.firstIndex(of: "("),
-           let closeIdx = trimmed.firstIndex(of: ")"),
-           openIdx < closeIdx {
-            let hint = trimmed[trimmed.index(after: openIdx)..<closeIdx]
-            let base = trimmed[..<openIdx]
-            self.parentheticalHint = hint.trimmingCharacters(in: .whitespaces).lowercased()
-            self.normalized = base.trimmingCharacters(in: .whitespaces).lowercased()
-        } else {
-            self.parentheticalHint = nil
-            self.normalized = trimmed.lowercased()
-        }
-    }
-}
-
-/// Maps semantic roles to column indexes. Built once per header row and
-/// reused by the per-row parsers.
-///
-/// Ambiguity handling: if a header matches multiple roles (e.g. `type`
-/// → `.measurementType` and `.scanType`), both entries point to the
-/// same column index. Format detection resolves which one is load-bearing.
-struct HeaderMap {
-    /// Headers as they appeared in the file (for error messages).
-    let rawHeaders: [String]
-    /// Role → column index, first match wins.
-    let roleIndex: [ColumnRole: Int]
-    /// Role → parenthesized hint captured from that column's header,
-    /// if any (e.g. `"lb"` from `Weight (lb)`).
-    let roleHint: [ColumnRole: String]
-
-    static func build(from headers: [String]) -> HeaderMap {
-        var roleIndex: [ColumnRole: Int] = [:]
-        var roleHint: [ColumnRole: String] = [:]
-
-        for (columnIndex, raw) in headers.enumerated() {
-            let norm = NormalizedHeader(raw)
-            for role in ColumnSynonyms.roles(forNormalized: norm.normalized) where roleIndex[role] == nil {
-                roleIndex[role] = columnIndex
-                if let hint = norm.parentheticalHint {
-                    roleHint[role] = hint
-                }
-            }
-        }
-
-        return HeaderMap(
-            rawHeaders: headers,
-            roleIndex: roleIndex,
-            roleHint: roleHint
-        )
-    }
-
-    func has(_ role: ColumnRole) -> Bool { roleIndex[role] != nil }
-
-    func hasAll(_ roles: Set<ColumnRole>) -> Bool {
-        roles.allSatisfy { has($0) }
-    }
-
-    func hasAny(_ roles: Set<ColumnRole>) -> Bool {
-        roles.contains { has($0) }
-    }
-
-    /// Which of the given roles are missing. Used to build user-facing
-    /// error messages when required columns aren't satisfied.
-    func missing(_ roles: Set<ColumnRole>) -> [ColumnRole] {
-        roles.filter { !has($0) }.sorted { $0.rawValue < $1.rawValue }
-    }
-
-    /// Raw value at the column matching `role`, trimmed. Returns nil if
-    /// the role is unmapped, the row is short, or the cell is blank.
-    func value(_ columns: [String], for role: ColumnRole) -> String? {
-        guard let index = roleIndex[role], index < columns.count else { return nil }
-        let v = columns[index].trimmingCharacters(in: .whitespaces)
-        return v.isEmpty ? nil : v
-    }
-
-    /// Parenthesized hint captured from `role`'s header cell, if any.
-    func hint(for role: ColumnRole) -> String? { roleHint[role] }
-}
-
-// MARK: - Date parsing
-
-/// Tries a fixed list of format strings in order until one matches. When
-/// a separate time column is also provided, each base format is tried
-/// alongside `HH:mm:ss` and `HH:mm` variants.
-///
-/// **Extension point:** to accept a new date format, add its
-/// `DateFormatter.dateFormat` string to `baseFormats`. Preformatted
-/// `DateFormatter` instances are cached at load time (POSIX locale,
-/// `twoDigitStartDate` anchored to year 2000) so parsing 10,000+ rows
-/// doesn't allocate a formatter per row.
-enum FlexibleDateParser {
-    /// Format strings for the date portion. Ordered so that 2-digit-year
-    /// variants are tried BEFORE 4-digit — otherwise DateFormatter in
-    /// lenient mode will parse "5/26/21" as the year 0021. We also set
-    /// `isLenient = false` below to make format boundaries strict.
-    private static let baseFormats: [String] = [
-        // ISO 8601 variants
-        "yyyy-MM-dd'T'HH:mm:ss.SSSXXXXX",
-        "yyyy-MM-dd'T'HH:mm:ssXXXXX",
-        "yyyy-MM-dd'T'HH:mm:ss",
-        "yyyy-MM-dd HH:mm:ss",
-        "yyyy-MM-dd",
-
-        // Slash-separated — 2-digit year first to anchor into the 2000s
-        // before the 4-digit patterns have a chance to match loosely.
-        "MM/dd/yy",
-        "M/d/yy",
-        "MM/dd/yyyy",
-        "M/d/yyyy",
-
-        // Other
-        "yyyy/MM/dd",
-        "dd-MM-yyyy"
-    ]
-
-    /// Pre-built formatters covering each base format plus two common
-    /// time-column append variants. Order matches `baseFormats`.
-    private static let formatters: [DateFormatter] = {
-        // `Calendar.date(from:)` requires enough components to pin down a
-        // unique instant — year alone returns nil. Supply month + day so
-        // twoDigitStartDate is actually set and `yy` patterns anchor to
-        // the 2000s instead of falling back to DateFormatter's default
-        // sliding window (which can wander into year 0001 CE).
-        let anchorYear2000: Date? = {
-            var cal = Calendar(identifier: .gregorian)
-            cal.locale = Locale(identifier: "en_US_POSIX")
-            return cal.date(from: DateComponents(year: 2000, month: 1, day: 1))
-        }()
-
-        return baseFormats.flatMap { base -> [DateFormatter] in
-            [base, "\(base) HH:mm:ss", "\(base) HH:mm"].map { pattern in
-                let f = DateFormatter()
-                f.calendar = Calendar(identifier: .gregorian)
-                f.locale = Locale(identifier: "en_US_POSIX")
-                f.timeZone = TimeZone.current
-                f.dateFormat = pattern
-                f.twoDigitStartDate = anchorYear2000
-                f.isLenient = false
-                return f
-            }
-        }
-    }()
-
-    /// Parses a date string (optionally combined with a separate time
-    /// column) against every registered format, returning the first
-    /// successful match whose year is plausibly modern. Returns nil if
-    /// nothing parses, or if every match lands outside 1900..3000 (which
-    /// indicates a `yyyy` pattern accepted a 2-digit year literally).
-    static func parse(date dateString: String, time timeString: String? = nil) -> Date? {
-        let dateTrim = dateString.trimmingCharacters(in: .whitespaces)
-        guard !dateTrim.isEmpty else { return nil }
-
-        let combined: String
-        if let timeString, !timeString.trimmingCharacters(in: .whitespaces).isEmpty {
-            combined = "\(dateTrim) \(timeString.trimmingCharacters(in: .whitespaces))"
-        } else {
-            combined = dateTrim
-        }
-
-        // Calendar for year validation — same gregorian/POSIX as parsers.
-        var validationCal = Calendar(identifier: .gregorian)
-        validationCal.locale = Locale(identifier: "en_US_POSIX")
-        validationCal.timeZone = TimeZone.current
-
-        for formatter in formatters {
-            guard let candidate = formatter.date(from: combined) else { continue }
-            let year = validationCal.component(.year, from: candidate)
-            if year >= 1900 && year <= 3000 {
-                return candidate
-            }
-            // Otherwise this formatter produced an obviously-wrong year
-            // (e.g. `yyyy` pattern gobbled a 2-digit year as year 21).
-            // Skip it and let the next format attempt a real match.
-        }
-        return nil
-    }
-}
-
-// MARK: - Unit resolution
-
-/// Resolves the canonical unit for a weight row from, in priority order:
-/// 1. An explicit unit column cell value
-/// 2. A parenthesized hint in the weight column's header (`Weight (kg)`)
-/// 3. The supplied default (usually the user's app-wide preference)
-///
-/// Returns nil if no source yields a recognisable unit — callers decide
-/// whether to fall back further or reject the row.
-enum WeightUnitResolver {
-    static func resolve(
-        explicit: String?,
-        headerHint: String?,
-        default defaultUnit: String
-    ) -> String? {
-        if let normalized = normalize(explicit) { return normalized }
-        if let normalized = normalize(headerHint) { return normalized }
-        if let normalized = normalize(defaultUnit) { return normalized }
-        return nil
-    }
-
-    private static func normalize(_ raw: String?) -> String? {
-        guard let raw = raw?.trimmingCharacters(in: .whitespaces), !raw.isEmpty else {
-            return nil
-        }
-        switch raw.lowercased() {
-        case "lb", "lbs", "pound", "pounds": return "lb"
-        case "kg", "kgs", "kilogram", "kilograms": return "kg"
-        default: return nil
-        }
-    }
-}
-
-/// Resolves the canonical length unit for a measurement row and converts
-/// raw numeric values to centimetres (Baseline's storage unit). Falls
-/// back to the supplied default if neither the column nor the header
-/// identifies a unit.
-enum LengthUnitResolver {
-    static func resolveUnit(
-        explicit: String?,
-        headerHint: String?,
-        default defaultUnit: String
-    ) -> String? {
-        if let n = normalize(explicit) { return n }
-        if let n = normalize(headerHint) { return n }
-        return normalize(defaultUnit)
-    }
-
-    static func toCentimeters(_ value: Double, unit: String) -> Double {
-        switch unit {
-        case "cm": return value
-        case "in": return value * 2.54
-        default: return value
-        }
-    }
-
-    private static func normalize(_ raw: String?) -> String? {
-        guard let raw = raw?.trimmingCharacters(in: .whitespaces), !raw.isEmpty else {
-            return nil
-        }
-        switch raw.lowercased() {
-        case "cm", "centimeter", "centimeters", "centimetres": return "cm"
-        case "in", "inch", "inches": return "in"
-        default: return nil
-        }
-    }
-}
-
-// MARK: - Format detection
-
-extension CSVFormat {
-    /// Top-level entry point: strip BOM + normalise line endings, extract
-    /// the header, and hand it to `detect(from: HeaderMap)`.
-    static func detect(from csv: String) -> CSVFormat? {
-        let cleaned = csv.stripBOM()
-            .replacingOccurrences(of: "\r\n", with: "\n")
-            .replacingOccurrences(of: "\r", with: "\n")
-        let firstLine = cleaned.split(separator: "\n", maxSplits: 1).first.map(String.init) ?? ""
-
-        // Parse the first line through the full quoting-aware parser so
-        // quoted headers (`"Weight (lb)"`) resolve the same way values do.
-        let lines = CSVImporter.parseLines(firstLine)
-        guard let headers = lines.first else { return nil }
-        return detect(from: HeaderMap.build(from: headers))
-    }
-
-    /// Role-based detection. Most-specific format wins so a scan CSV
-    /// (which includes a weight column) isn't misclassified as weights.
-    static func detect(from headerMap: HeaderMap) -> CSVFormat? {
-        let scanCoreFields: Set<ColumnRole> = [
-            .scanWeightKg, .scanSMM, .scanBFM, .scanPBF, .scanTBW, .scanBMI, .scanBMR
-        ]
-        if headerMap.has(.date) && headerMap.hasAll(scanCoreFields) {
-            return .scans
-        }
-
-        let hasMeasurementType = headerMap.hasAny([.measurementType, .scanType])
-        if headerMap.has(.date) && hasMeasurementType && headerMap.has(.measurementValue) {
-            return .measurements
-        }
-
-        if headerMap.has(.date) && headerMap.has(.weight) {
-            return .weights
-        }
-
-        return nil
-    }
-}
-
-// MARK: - String helpers
-
-private extension String {
-    /// Removes a leading UTF-8 BOM (`U+FEFF`) if present. Excel writes one
-    /// when exporting CSVs on macOS — silently breaking header matching.
-    func stripBOM() -> String {
-        guard hasPrefix("\u{FEFF}") else { return self }
-        return String(dropFirst())
-    }
-}
-
 // MARK: - Importer
 
 enum CSVImporter {
